Rewrite to remap the -1 map_index TI if possible

+++ airflow/models/abstractoperator.py
@@ -21,6 +21,8 @@ import datetime
 import inspect
 from typing import TYPE_CHECKING, Any, Callable, ClassVar, Collection, Iterable, Iterator, Sequence
 
+from sqlalchemy.sql import exists
+
 from airflow.compat.functools import cache, cached_property
 from airflow.configuration import conf
 from airflow.exceptions import AirflowException
@@ -494,10 +496,25 @@ class AbstractOperator(LoggingMixin, DAGNode):
                 )
                 unmapped_ti.state = TaskInstanceState.SKIPPED
             else:
-                # Delete the original task so we can create new ones
-                self.log.debug("Deleting the original task instance: %s", unmapped_ti)
+                zero_index_ti_exists = session.query(
+                    exists().where(
+                        TaskInstance.dag_id == self.dag_id,
+                        TaskInstance.task_id == self.task_id,
+                        TaskInstance.run_id == run_id,
+                        TaskInstance.map_index == 0,
+                    )
+                ).scalar()
+                if not zero_index_ti_exists:
+                    # Otherwise convert this into the first mapped index, and create
+                    # TaskInstance for other indexes.
+                    unmapped_ti.map_index = 0
+                    self.log.debug("Updated in place to become %s", unmapped_ti)
+                    all_expanded_tis.append(unmapped_ti)
+                    session.flush()
+                else:
+                    self.log.debug("Deleting the original task instance: %s", unmapped_ti)
+                    session.delete(unmapped_ti)
                 state = unmapped_ti.state
-                session.delete(unmapped_ti)
 
         if total_length is None or total_length < 1:
             # Nothing to fixup.
